Reorganise file layout, reduce duplicated code.
