+++ martini/mochi/_dtypes.py
@@ -1,0 +1,1 @@
+CELL_DTYPE = [("x", float), ("y", float), ("z", float), ("size", float)]
